更新前端代码

+++ src/page/information/index.tsx
@@ -1,99 +1,103 @@
 import './index.less';
-import { Calendar } from 'antd';
-import type { CalendarMode } from 'antd/es/calendar/generateCalendar';
-import type { Moment } from 'moment';
-import React from 'react';
-import icon_yoga from '../../../public/img/class/yoga.svg';
+import React, { useEffect } from 'react';
+import icon_yoga from '../../../public/img/class/瑜伽.svg';
 import { useState } from 'react';
-import {
-  Alert,
-  Button,
-  Modal,
-  TimePicker,
-  Form,
-  Input,
-  InputNumber,
-  Select,
-  Row,
-  Col,
-} from 'antd';
-import { LeftCircleTwoTone } from '@ant-design/icons';
+import { Alert, Row, Col } from 'antd';
+import { LeftOutlined } from '@ant-design/icons';
 import SlideItem from '../../component/slideItem';
-
-const c_name = new URLSearchParams(window.location.search).get('c_name');
-const date = new URLSearchParams(window.location.search).get('date');
-console.log(date, c_name);
+import moment from 'moment';
+import Api from '@/request/request';
 
 const Information: React.FC = () => {
-  const [item, setItem] = useState({
+  const [m_item, setM_item] = useState({
+    c_id: 1,
     c_name: '瑜伽',
-    date: '2022-07-19',
-    s_time: '08:00',
-    e_time: '09:00',
-    place: '瑜伽馆',
-    price_s: '100',
-    price: '200',
+    s_time: moment(),
+    time_long: '18:30:00',
+    place: '师生活动中心2-208',
+    price: 400,
     num: 10,
-    img: icon_yoga,
   });
+  // 从url中获得get传值
+  const get = (name: string) => {
+    // 从url中获得get传值
+    let arg: any = new URLSearchParams(window.location.search).get(name);
+    return arg;
+  };
+  m_item.c_id = get('c_id');
+  m_item.c_name = get('c_name');
+  m_item.s_time = moment(get('s_time'), 'YYYY-MM-DD HH:mm');
+  m_item.time_long = get('time_long');
+  m_item.place = get('place');
+  m_item.price = get('price');
+  m_item.num = get('num');
+  let s_price = (
+    m_item.num < 5 ? m_item.price / 5 : m_item.price / m_item.num
+  ).toFixed(2);
+  let l_price = (
+    m_item.num < 5
+      ? (m_item.price / 5) * 1.5
+      : (m_item.price / m_item.num) * 1.5
+  ).toFixed(2);
+  console.log(m_item);
+
   const [s_info, setS_info] = useState([
     {
-      s_name: '张3',
-      sign_up: '2022-07-19',
-      pay: 100,
-      default: false,
-    },
-    {
-      s_name: '张4',
-      sign_up: '2022-07-19',
-      pay: 100,
-      default: false,
-    },
-    {
-      s_name: '张5',
-      sign_up: '2022-07-19',
-      pay: 100,
-      default: false,
-    },
-    {
-      s_name: '张6',
-      sign_up: '2022-07-19',
-      pay: 100,
-      default: false,
-    },
-    {
-      s_name: '张7',
-      sign_up: '2022-07-19',
-      pay: 100,
-      default: false,
-    },
-    {
-      s_name: '李四',
-      sign_up: '2022-07-21',
-      pay: 200,
-      default: false,
-    },
-    {
-      s_name: '李四',
-      sign_up: '2022-07-21',
-      pay: 200,
-      default: false,
-    },
-    {
-      s_name: '李四',
-      sign_up: '2022-07-21',
-      pay: 200,
+      c_id: 1,
+      c_name: '瑜伽',
+      id: 4,
+      pay: 0,
+      signup: '08-06 10:00',
+      u_id: '2020212205248',
+      u_name: 'lcy4',
       default: true,
     },
   ]);
+  const api = new Api();
+  useEffect(() => {
+    api
+      .GetSignupUsers({
+        params: {
+          c_id: m_item.c_id,
+        },
+      })
+      .then((res: any) => {
+        if (res.code === 1) {
+          console.log('success', res);
+          setS_info(res.data);
+        } else {
+          console.log('error', res);
+        }
+      })
+      .catch((err: any) => {
+        console.log(err);
+      });
+  }, []);
+  console.log(s_info);
 
   const onDelete = (value: any) => {
     return () => {
       console.log(value);
       setS_info(
         s_info.map((item: any) => {
-          if (item.s_name === value.s_name) {
+          if (item.u_name === value.u_name) {
             item.default = true;
+            // 黑名单添加
+            api
+              .AddDefault({
+                c_name: m_item.c_name,
+                u_id: item.u_id,
+              })
+              .then((res: any) => {
+                if (res.code === 1) {
+                  console.log('success', res);
+                } else {
+                  console.log('error', res);
+                }
+              })
+              .catch((err: any) => {
+                console.log(err);
+              });
           }
           return item;
         }),
@@ -108,22 +112,35 @@ const Information: React.FC = () => {
   return (
     <div className="information">
       <div className="title">
-        <LeftCircleTwoTone onClick={backHome} />
+        <LeftOutlined onClick={backHome} />
         <span>课程详情</span>
       </div>
       <div className="m-alert">
         <Alert
-          message={item.s_time + '~' + item.e_time + ' ' + item.place}
+          message={
+            moment(m_item.s_time).format('HH:mm') +
+            '~' +
+            moment(m_item.s_time).add(m_item.time_long, 'M').format('HH:mm') +
+            ' ' +
+            m_item.place
+          }
           type="info"
-          icon={<img src={item.img} alt="" />}
+          icon={
+            <img
+              src={require('../../../public/img/class/' +
+                m_item.c_name +
+                '.svg')}
+              alt=""
+            />
+          }
           showIcon
         />
       </div>
       <div className="m-title">
         <Row>
-          <Col span={8}>{'预约：' + item.price_s}</Col>
-          <Col span={8}>{'非预约：' + item.price}</Col>
-          <Col span={8}>{'预约人数：' + item.num}</Col>
+          <Col span={8}>{'预约：' + s_price}</Col>
+          <Col span={8}>{'非预约：' + l_price}</Col>
+          <Col span={8}>{'预约人数：' + m_item.num}</Col>
         </Row>
       </div>
       <div className="m-cnt">
@@ -136,20 +153,31 @@ const Information: React.FC = () => {
         </div>
         <div className="m-cnt-list">
           {s_info.map((item, index) => {
-            if (item.default === false) {
+            if (
+              item.default === false &&
+              m_item.s_time < moment() &&
+              localStorage.getItem('u_id') != null
+            ) {
               return (
                 <SlideItem
                   key={index}
                   children={
                     <Row
                       className={
-                        (date === item.sign_up ? 'today ' : '') +
-                        (item.default ? 'default' : '')
+                        (m_item.s_time.format('YYYY-MM-DD') === item.signup
+                          ? 'today '
+                          : '') + (item.default ? 'default' : '')
                       }
                     >
-                      <Col span={6}>{item.s_name}</Col>
-                      <Col span={12}>{item.sign_up}</Col>
-                      <Col span={6}>{item.pay}</Col>
+                      <Col span={6}>{item.u_name}</Col>
+                      <Col span={12}>
+                        {moment(item.signup).format('MM-DD HH:mm')}
+                      </Col>
+                      <Col span={6}>
+                        {m_item.s_time.format('YYYY-MM-DD') === item.signup
+                          ? l_price
+                          : s_price}
+                      </Col>
                     </Row>
                   }
                   onDelete={onDelete(item)}
@@ -160,13 +188,20 @@ const Information: React.FC = () => {
                 <Row
                   key={index}
                   className={
-                    (date === item.sign_up ? 'today ' : '') +
-                    (item.default ? 'default' : '')
+                    (m_item.s_time.format('YYYY-MM-DD') === item.signup
+                      ? 'today '
+                      : '') + (item.default ? 'default' : '')
                   }
                 >
-                  <Col span={6}>{item.s_name}</Col>
-                  <Col span={12}>{item.sign_up}</Col>
-                  <Col span={6}>{item.pay}</Col>
+                  <Col span={6}>{item.u_name}</Col>
+                  <Col span={12}>
+                    {moment(item.signup).format('MM-DD HH:mm')}
+                  </Col>
+                  <Col span={6}>
+                    {m_item.s_time.format('YYYY-MM-DD') === item.signup
+                      ? l_price
+                      : s_price}
+                  </Col>
                 </Row>
               );
             }
